Add base classes for org level views
