+++ temba/tickets/views.py
@@ -21,7 +21,8 @@ from django.utils.translation import gettext_lazy as _
 from temba.msgs.models import Msg
 from temba.notifications.views import NotificationTargetMixin
 from temba.orgs.mixins import OrgObjPermsMixin, OrgPermsMixin
-from temba.orgs.views import BaseExportView, BaseMenuView, ModalMixin
+from temba.orgs.views import BaseExportView, ModalMixin
+from temba.orgs.views_base import BaseMenuView
 from temba.utils.dates import datetime_to_timestamp, timestamp_to_datetime
 from temba.utils.export import response_from_workbook
 from temba.utils.fields import InputWidget
